refactor: add package for testing cli
