+++ clint/cli/__init__.py
@@ -1,0 +1,3 @@
+"""Command line interface classes."""
+
+from .command import Command
